Consolidate MatchViewController and ProfileViewController
