+++ Pear/Controller/CommunityViewController.swift
@@ -114,14 +114,8 @@ extension CommunityViewController: UITableViewDataSource {
 
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         let user = users[indexPath.row]
-        navigationController?.pushViewController(
-            ProfileViewController(
-                userId: user.netID,
-                userName: "\(user.firstName ?? "") \(user.lastName ?? "")",
-                profilePicture: user.profilePictureURL ?? ""
-            ), animated: true)
+        navigationController?.pushViewController(ProfileViewController(match: nil, user: nil, userId: user.netID, type: .user), animated: true)
     }
-
 }
 
 // MARK: - SearchBarDelegate
